Finalize PGN download API and openings API refactor with modularized backend structure. Verified passing automation tests.
